+++ crates/toml_datetime/CHANGELOG.md
@@ -1,0 +1,11 @@
+# Changelog
+
+The format is based on [Keep a Changelog].
+
+[Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
+
+<!-- next-header -->
+## [Unreleased] - ReleaseDate
+
+<!-- next-url -->
+[Unreleased]: https://github.com/toml-rs/toml_edit/compare/87741642c0f1a5217fd125e99fb52181869f74fa...HEAD
