+++ crates/toml_edit/Cargo.toml
@@ -36,7 +36,8 @@ pre-release-replacements = [
 [features]
 default = []
 easy = ["serde"]
-perf = ["kstring"]
+perf = ["dep:kstring"]
+serde = ["dep:serde", "toml_datetime/serde"]
 
 [dependencies]
 indexmap = "1.9.1"
@@ -44,6 +45,7 @@ combine = "4.6.6"
 itertools = "0.10.5"
 serde = { version = "1.0.145", features = ["derive"], optional = true }
 kstring = { version = "2.0.0", features = ["max_inline"], optional = true }
+toml_datetime = { version = "0.5.0", path = "../toml_datetime" }
 
 [dev-dependencies]
 serde_json = "1.0.85"
