+++ crates/toml_edit/src/easy/datetime.rs
@@ -1,1 +1,1 @@
-pub use crate::datetime::*;
+pub use toml_datetime::*;
