+++ crates/toml_edit/src/easy/value.rs
@@ -521,7 +521,7 @@ impl<'a, 'de> serde::de::Visitor<'de> for DatetimeOrTable<'a> {
     where
         E: serde::de::Error,
     {
-        if s == crate::datetime::dt_serde::FIELD {
+        if s == toml_datetime::__unstable::FIELD {
             Ok(true)
         } else {
             self.key.push_str(s);
@@ -533,7 +533,7 @@ impl<'a, 'de> serde::de::Visitor<'de> for DatetimeOrTable<'a> {
     where
         E: serde::de::Error,
     {
-        if s == crate::datetime::dt_serde::FIELD {
+        if s == toml_datetime::__unstable::FIELD {
             Ok(true)
         } else {
             *self.key = s;
