+++ crates/toml_edit/src/parser/datetime.rs
@@ -1,10 +1,11 @@
-use crate::datetime::*;
-use crate::parser::errors::CustomError;
-use crate::parser::trivia::from_utf8_unchecked;
 use combine::parser::byte::byte;
 use combine::parser::range::{recognize, take_while1};
 use combine::stream::RangeStream;
 use combine::*;
+use toml_datetime::*;
+
+use crate::parser::errors::CustomError;
+use crate::parser::trivia::from_utf8_unchecked;
 
 // ;; Date and Time (as defined in RFC 3339)
